requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.16.1
pymongo==4.6.0
Flask==2.3.5
flask-cors==4.0.0